+++ NEWS.md
@@ -1,3 +1,13 @@
+# missRanger 2.4.0
+
+## Enhancements
+
+- New argument `data_only = TRUE` to control if only the imputed data should returned (default), or an object of class "missRanger". This object contains the imputed data and infos like OOB prediction errors. This fixes [issue 28](https://github.com/mayer79/missRanger/issues/28). The value `FALSE` will become the default in {missRanger 3.0.0}. This will be announced via deprecation cycle.
+
+## Bug fixes
+
+- In case the algorithm did not converge, the data of the *last* iteration was returned instead of the current one. This has been fixed.
+
 # missRanger 2.3.0
 
 ## Major improvements